+++ source/DSCResources/MSFT_DhcpServerOptionValue/MSFT_DhcpServerOptionValue.psm1
@@ -98,7 +98,7 @@ function Set-TargetResource
         $OptionId,
 
         [Parameter()]
-        [String[]]
+        [System.String[]]
         $Value,
 
         [Parameter(Mandatory = $true)]
@@ -161,7 +161,7 @@ function Test-TargetResource
         $OptionId,
 
         [Parameter()]
-        [String[]]
+        [System.String[]]
         $Value,
 
         [Parameter(Mandatory = $true)]